+++ tests/test_queries/test_annotations.py
@@ -4,6 +4,7 @@ import pytest
 from django import VERSION as DJANGO_VERSION
 from django.db import connection, models
 
+from queryable_properties.query import QUERYING_PROPERTIES_MARKER
 from queryable_properties.utils.internal import get_queryable_property_descriptor
 from ..app_management.models import (
     ApplicationWithClassBasedProperties, ApplicationWithDecoratorBasedProperties, VersionWithClassBasedProperties,
@@ -28,10 +29,12 @@ class TestAggregateAnnotations(object):
         queryset = model.objects.filter(**filters).select_properties('version_count', 'major_sum').filter(**filters)
         assert 'version_count' in queryset.query.annotations
         assert 'major_sum' in queryset.query.annotations
-        assert all(model.version_count.has_cached_value(obj) for obj in queryset)
-        assert all(obj.version_count == 4 for obj in queryset)
-        assert all(model.major_sum.has_cached_value(obj) for obj in queryset)
-        assert all(obj.major_sum == 5 for obj in queryset)
+        for application in queryset:
+            assert model.version_count.has_cached_value(application)
+            assert application.version_count == 4
+            assert model.major_sum.has_cached_value(application)
+            assert application.major_sum == 5
+            assert not hasattr(application, QUERYING_PROPERTIES_MARKER)
 
     @pytest.mark.parametrize('model, limit, expected_total', [
         (ApplicationWithClassBasedProperties, None, 8),
@@ -98,6 +101,21 @@ class TestAggregateAnnotations(object):
         del queryset.query.annotations['version_count']
         assert bool(queryset)
         assert all(not model.version_count.has_cached_value(obj) for obj in queryset)
+
+    @pytest.mark.skipif(DJANGO_VERSION < (1, 7), reason="Raw queries didn't exist before Django 1.7")
+    @pytest.mark.parametrize('model', [ApplicationWithClassBasedProperties, ApplicationWithDecoratorBasedProperties])
+    def test_raw_query(self, model):
+        pks, names = zip(*model.objects.values_list('pk', 'name'))
+        queryset = model.objects.raw('SELECT id, name, 5 AS version_count FROM {}'.format(model._meta.db_table))
+        counter = 0
+        for application in queryset:
+            assert model.version_count.has_cached_value(application) is True
+            assert application.version_count == 5
+            assert application.pk in pks
+            assert application.name in names
+            assert not hasattr(application, QUERYING_PROPERTIES_MARKER)
+            counter += 1
+        assert counter == 2
 
 
 @pytest.mark.skipif(DJANGO_VERSION < (1, 8), reason="Expression-based annotations didn't exist before Django 1.8")